# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Dubhu Designer API
  description: API specification for Dubhu Designer services.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
...
  /concierges/{conciergeId}/agents:
    post:
      summary: Create a new agent
      description: Creates a new agent specification scoped to the specified concierge. Requires name, description, goal, and model. Agent name must be unique within the same concierge.
      operationId: createAgent
      tags:
        - Agents
      parameters:
        - name: conciergeId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the parent concierge
          schema:
            type: string
            example: "651da3b57f8641a27e36c579"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
            example:
              name: "research_assistant"
              description: "Conducts deep literature research and analysis."
              goal: "Summarize research papers and extract key insights."
              model: "gemini-3.5-flash"
      responses:
        '201':
          description: Agent created successfully
          headers:
            ETag:
              description: Strong ETag for the created agent representation
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Agent'
        '400':
          description: Bad request (invalid payload or validation failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Validation failed"
                details:
                  - field: "name"
                    reason: "name is required and cannot be blank"
                    value: ""
        '404':
          description: Concierge not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Concierge not found"
        '409':
          description: Conflict (agent name already exists for this concierge)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Conflict"
                details:
                  - field: "name"
                    reason: "agent name already exists for this concierge"
                    value: "research_assistant"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "failed to insert agent"
    get:
      summary: List all agents for a concierge
      description: Retrieves a list of all agents belonging to the specified concierge.
      operationId: listAgents
      tags:
        - Agents
      parameters:
        - name: conciergeId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the parent concierge
          schema:
            type: string
            example: "651da3b57f8641a27e36c579"
      responses:
        '200':
          description: A list of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
        '404':
          description: Concierge not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Concierge not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "failed to query agents"
  /concierges/{conciergeId}/agents/{agentId}:
    get:
      summary: Get agent by ID
      description: Retrieves a specific agent by its MongoDB ObjectID under the specified concierge.
      operationId: getAgentById
      tags:
        - Agents
      parameters:
        - name: conciergeId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the parent concierge
          schema:
            type: string
            example: "651da3b57f8641a27e36c579"
        - name: agentId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the agent
          schema:
            type: string
            example: "651da3b57f8641a27e36c582"
      responses:
        '200':
          description: Agent retrieved successfully
          headers:
            ETag:
              description: Strong ETag for the current agent representation
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Agent'
        '404':
          description: Concierge or agent not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Agent not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "failed to find agent"
    put:
      summary: Update agent by ID
      description: Updates an existing agent specification under the specified concierge. Name, description, goal, and model can be updated. Requires the strong ETag returned when the agent was last read.
      operationId: updateAgent
      tags:
        - Agents
      parameters:
        - name: conciergeId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the parent concierge
          schema:
            type: string
            example: "651da3b57f8641a27e36c579"
        - name: agentId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the agent
          schema:
            type: string
            example: "651da3b57f8641a27e36c582"
        - name: If-Match
          in: header
          required: true
          description: Strong ETag from the latest agent representation
          schema:
            type: string
            example: '"1"'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
            example:
              name: "research_assistant_v2"
              description: "Updated deep literature research and analysis agent."
              goal: "Summarize research papers, extract key insights, and generate critical evaluations."
              model: "gemini-3.5-flash"
      responses:
        '200':
          description: Agent updated successfully
          headers:
            ETag:
              description: Strong ETag for the updated agent representation
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Agent'
        '400':
          description: Bad request (invalid payload or validation failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Validation failed"
                details:
                  - field: "name"
                    reason: "name cannot exceed 200 characters"
                    value: "..."
        '404':
          description: Concierge or agent not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Agent not found"
        '409':
          description: Conflict (agent name already exists for this concierge)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Conflict"
                details:
                  - field: "name"
                    reason: "agent name already exists for this concierge"
                    value: "research_assistant"
        '412':
          description: The supplied If-Match ETag does not match the current agent representation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '428':
          description: If-Match header is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "failed to update agent"
  /concierges/{conciergeId}/agents/{agentId}/instructions/{instructionId}:
    post:
      summary: Assign an instruction to an agent
      description: Assigns an existing instruction specification to the agent. Validates that the concierge, agent, and instruction exist. Duplicate assignments are handled idempotently.
      operationId: assignInstructionToAgent
      tags:
        - Agents
      parameters:
        - name: conciergeId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the parent concierge
          schema:
            type: string
            example: "651da3b57f8641a27e36c579"
        - name: agentId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the agent
          schema:
            type: string
            example: "651da3b57f8641a27e36c582"
        - name: instructionId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the instruction
          schema:
            type: string
            example: "651da3b57f8641a27e36c580"
      responses:
        '204':
          description: Instruction assigned successfully
        '404':
          description: Concierge, agent, or instruction not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Instruction not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "failed to assign instruction to agent"
    delete:
      summary: Unassign an instruction from an agent
      description: Removes an instruction assignment from the agent. Validates that the concierge and agent exist. If the instruction is not assigned, the operation succeeds idempotently.
      operationId: unassignInstructionFromAgent
      tags:
        - Agents
      parameters:
        - name: conciergeId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the parent concierge
          schema:
            type: string
            example: "651da3b57f8641a27e36c579"
        - name: agentId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the agent
          schema:
            type: string
            example: "651da3b57f8641a27e36c582"
        - name: instructionId
          in: path
          required: true
          description: Hexadecimal MongoDB ObjectID of the instruction
          schema:
            type: string
            example: "651da3b57f8641a27e36c580"
      responses:
        '204':
          description: Instruction unassigned successfully
        '404':
          description: Concierge, agent, or instruction not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Agent not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "failed to unassign instruction from agent"
components:
  schemas:
...
    CreateAgentRequest:
      type: object
      required:
        - name
        - description
        - goal
        - model
      properties:
        name:
          type: string
          maxLength: 200
          description: Unique agent name
          example: "research_assistant"
        description:
          type: string
          maxLength: 2000
          description: Purpose and capabilities of the agent
          example: "Conducts deep literature research and analysis."
        goal:
          type: string
          maxLength: 2000
          description: Primary objective or mission of the agent
          example: "Summarize research papers and extract key insights."
        model:
          type: string
          enum:
            - "gemini-3.5-flash"
            - "gemini-3.5-flash-lite"
          description: Supported foundation model identifier
          example: "gemini-3.5-flash"
    UpdateAgentRequest:
      type: object
      required:
        - name
        - description
        - goal
        - model
      properties:
        name:
          type: string
          maxLength: 200
          description: Unique agent name within the concierge
          example: "research_assistant_v2"
        description:
          type: string
          maxLength: 2000
          description: Purpose and capabilities of the agent
          example: "Updated deep literature research and analysis agent."
        goal:
          type: string
          maxLength: 2000
          description: Primary objective or mission of the agent
          example: "Summarize research papers, extract key insights, and generate critical evaluations."
        model:
          type: string
          enum:
            - "gemini-3.5-flash"
            - "gemini-3.5-flash-lite"
          description: Supported foundation model identifier
          example: "gemini-3.5-flash"
    Agent:
      type: object
      required:
        - id
        - conciergeId
        - name
        - description
        - goal
        - model
        - instructions
      properties:
        id:
          type: string
          description: Hexadecimal MongoDB ObjectID of the agent
          example: "651da3b57f8641a27e36c579"
        conciergeId:
          type: string
          description: Hexadecimal MongoDB ObjectID of the parent concierge
          example: "651da3b57f8641a27e36c581"
        name:
          type: string
          example: "research_assistant"
        description:
          type: string
          example: "Conducts deep literature research and analysis."
        goal:
          type: string
          example: "Summarize research papers and extract key insights."
        model:
          type: string
          enum:
            - "gemini-3.5-flash"
            - "gemini-3.5-flash-lite"
          example: "gemini-3.5-flash"
        instructions:
          type: array
          description: List of hexadecimal MongoDB ObjectIDs of instructions assigned to this agent
          items:
            type: string
            example: "651da3b57f8641a27e36c580"
    FieldError:
      type: object
      required:
        - field
        - reason
      properties:
        field:
          type: string
          description: Name or path of the field that failed validation
          example: "name"
        reason:
          type: string
          description: Reason why the field failed validation
          example: "tool name cannot exceed 200 characters"
        value:
          nullable: true
          description: Value that caused the validation failure
          example: "A very long tool name..."
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Validation failed"
        details:
          oneOf:
            - type: array
              items:
                $ref: '#/components/schemas/FieldError'
            - type: string
              example: "Key: 'CreateToolDTO.Name' Error:Field validation for 'Name' failed on the 'required' tag"